+++ packages/common/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @7ito/sketcherson-common
 
+## 0.3.16
+
+### Patch Changes
+
+- Emote update
+
 ## 0.3.15
 
 ### Patch Changes