+++ packages/server/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @7ito/sketcherson-server
 
+## 0.3.16
+
+### Patch Changes
+
+- Emote update
+- Updated dependencies
+  - @7ito/sketcherson-common@0.3.16
+
 ## 0.3.15
 
 ### Patch Changes